Move source code into src and examples into their own directory
