+++ examples/vue-app/main.js
