+++ examples/vue-app/public/favicon.ico
